move class into src and namespace
move class into SeaTable\SeaTableApi namespace and rename it to
SeaTableApi (API -> Api).

renaming the class alone would not be a problem on PHPs end (as class names
are case in-sensitive), however this is not compatible with composer auto-
loader which does binary string comparison for class mappings.

as introducing the namespace would break it on the PHP level anyway, this
is the perfect place to rename the class.

not only that, the original class-name is preserved for auto-loading in
Compat/Backward so that an upgrade from 0.0.x to 0.1.x won't break places
where the old class-name is in use with autoloading.