limit url schemes to http/s (url option)
previously the 'url' option accepted any protocol. this is cumbersome as
curl supports very many but the standard schemes with SeaTable API are
http and https (and not like file:/// which curl has no problem to open).

adding tests for scheme url regressions to cover the changes. as the code
is in the constructor, by checking for an error on url validation and the
good case is to get an error on the missing user.