backwards compatibility baseline in 0.x
give more general information about deprecation and the public interface
for very early versions to highlight the differences in semantic
versioning 2.0.0 as if doing stable releases.

in short: patch is minor and next patch can (but must not) be major.
normally minor is major.

that is a bit more precise than just saying in 0.x everything can break.

it also highlights limitations we see for deprecation messages.

finally the baseline configuration for patch bc breaks in roave
backwards-compatibility-check via roave-bc-with-config-support in
nyholm/roave-bc-check-ga as in use in the github action.