add base test-case
group all test-cases for forward/backward compatibility when phpunit
changes (or alternatively use symfony for that). also towards other
migrations.

the change in the version test of SeaTableApiTest is that composer - when
not on a git branch (git rebase, detached HEAD etc.) - is not dev-main
(composer does some easing here building that string) but 1.0.0+no-version.