improve cURL error exception message
the empty response is a common case to throw here, but actually it must
not mean that there was an actual curl error.

therefore show the cURL error number (even if zero) and quote the cURL
error message (to show there is none).

additionally add a notice to the message if the response was an actual
string but just empty.