improve phpunit configuration
in f6bebc2 (add base test-case, 2022-01-05) it has been forgotten to
make use of it in the PHP-test as well.

additionally allow test on tag version like v0.1.10 which was not
possible previously.

additionally change to phpunit.xml.dist configuration to allow local
phpunit configuration.

finally cover the out of bounds exception and mark internal lines.

ref: f6bebc2ed17ab12de6ad92606b2789da8118e42e