fix error expecting tests
previously a test that expected an error did pass in case of no error.

reason: when introducing an error formulating the message as the test
fixture, an error was always given resulting in a false positive.

to improve, expect concrete error classes from phpunit, e.g. expectNotice()
or expectDeprecation() instead of just expectError().

additionally formulate the message fixture first, then expect the error. in
this order, the test will fail first as at the moment a message expectation
is done with errors, the test already fails as no error is expected yet.

this is not so much an issue with exceptions as normally exception classes
differentiate more.  however it is certainly good practice to first
formulate the exception message expectations before specifying the type of
exception.  note, this commit is about errors - not exceptions:

define error message expectations before expecting errors and precisely
define which class of errors are expected (notice, deprecation etc.) not
just "error".