method / parameter naming for 0.2.x
names and types of parameters and object return type hint.

check prefixes, verbs and words of the main class api methods as well as
parameter names and types.

to cover v0.2 changes with a test-case (naming/style):

- check for "A" (like AView) and make it a failure
- check for "_" (underscore) in parameter names
- check for "DTable" regressions in method names
- check for pagination parameter order
- check for method return type
- check for parameter type
- check for doc-comment
- check for doc-comment @link
- check for doc-comment @deprecated
- check for doc-comment @group (required for non-deprecated)

docs: upgrading (pagination parameters)

document the backwards incompatible change in 0.2.0 with the change of
order of the pagination parameters ($per_page, $page -> $page, $perPage).