improve method deprecation message
for method deprecation, having the location where a deprecated method call
is made at the end of a line, and then on the next line both first the call
point and the (as it can't be prevented) trigger_error() location for easy
tracking.

add new line:

    PHP Deprecated:  At in <path> on <line> and triggered in <path> on <line>

two spaces after Deprecated:__ intended, to align with PHP log format.