build on gh-actions with mkdocs + mike
gh-pages branch needs to be prepared for mike as this changes from single
commit to multiple on gh-pages branch.

git-fetch in CI is necessary, too, to preserve history of older versions.

especially prepared w/ main symbolic link which needs manual branch
patching. the procedure is to *not* use mike version aliases for symbolic
link "aliases".

mike creates empty commits if there are no changes.  the more determinitic
the build w/ mike is, the easier is it to maintain the gh-pages branch.

[Mike CI Docs]: https://github.com/jimporter/mike#deploying-via-ci