migrate symlink to main
main is the main version (stable). earlier we thought to name it latest as
that is the default alias in mkdocs/mike but we're not using aliases any
longer and want to map this to true branch names.

therefore main is the stable release branch and new versions can be added
as-is.

404.html needs fixing still thought.