fix list-daily-active-users date encoding
the in 0.1.21 deprecated method listDailyActiveUsers() passed the $date
parameter to its replacement method sysAdminListDailyActiveUsers()
unchanged while the replacement method contained a fix correctly url
encoding it.

fix is to decode it before passing it along.

the default value in the replacement method also was prone to the same
encoding flaw resulting in a double encoding when using the default
value.