shorten overlong results in messages
when throwing the internal response exception which often is in the 5xx and
4xx range a SeaTable may generate large responses due to server and/or
gateway configuration.  such responses are not displayed in full any longer
in the exception messages themselves but are shortened to 256 bytes.

since e6ba634 (add internal exceptions, 2022-09-01) the exception still has
the response in full via the internal getResult() method.

rationale is to effectively prevent overlong messages in the php error log
which is capped at 1024 bytes per message, or similar the nginx error log
which is capped at 2084 bytes per message (and it decorates the PHP SAPI
error messages w/ additional information).  the problem with the cut-off
is that often the stack-trace is then shortened or non-existent which
hinders analysis.